+++ eveuniverse/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "eveuniverse.apps.EveuniverseConfig"
 
-__version__ = "0.3.5"
+__version__ = "0.4.0"
 __title__ = "Eve Universe"